Rename conversion module to dangerzone-insecure-converter
