+++ Dockerfile
@@ -62,7 +62,9 @@ RUN mkdir -p /opt/dangerzone/dangerzone
 RUN touch /opt/dangerzone/dangerzone/__init__.py
 
 # Copy only the Python code, and not any produced .pyc files.
-COPY conversion/*.py /opt/dangerzone/dangerzone/conversion/
+# NOTE: Copy the code under `conversion/`, to mimic the old path where
+# Dangerzone looked for files.
+COPY dangerzone_insecure_converter/*.py /opt/dangerzone/dangerzone/conversion/
 
 # Create a directory that will be used by gVisor as the place where it will
 # store the state of its containers.
