+++ src/dangerzone_insecure_converter/doc_to_pixels.py
