+++ src/dangerzone_insecure_converter/errors.py
